+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/App.xaml.cs
@@ -1,6 +1,7 @@
-﻿using System.Configuration;
-using System.Data;
+﻿using System.Globalization;
 using System.Windows;
+using TD.WPF.Models;
+using TD.WPF.Views;
 
 namespace TD.WPF
 {
@@ -9,6 +10,18 @@ namespace TD.WPF
     /// </summary>
     public partial class App : Application
     {
-    }
+        private void OnStartup(object sender, StartupEventArgs e)
+        {
+            var settings = AppSettings.GetAppSettings();
 
+            if (!string.IsNullOrWhiteSpace(settings.UiCulture))
+            {
+                var culture = new CultureInfo(settings.UiCulture);
+                Thread.CurrentThread.CurrentUICulture = culture;
+                Thread.CurrentThread.CurrentCulture = culture;
+            }
+
+            new MainWindow().Show();
+        }
+    }
 }